backend/requirements.txt
fastapi>=0.100.0,<0.101.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
asyncmy>=0.2.0,<0.3.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv
aiomysql>=0.1.0,<0.2.0
email-validator>=2.0.0,<3.0.0
python-multipart>=0.0.5,<0.1.0
jinja2>=3.0.0,<4.0.0
aiosmtplib>=2.0.0,<3.0.0 